add event repo

+++ src/main/java/com/org/entities/Event.java
@@ -5,24 +5,33 @@
  */
 package com.org.entities;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+
 
 /**
  *
  * @author DEEV
  */
 @Entity
-@Table(name = "events")
+@Table(name= "events")
 public class Event {
     
   @Id
   @GeneratedValue(strategy=GenerationType.AUTO)
-  private int id;
+  private long id;
 
   @Column(name = "title")
   private String name;
@@ -36,4 +45,8 @@ public class Event {
   private String hour;
   @Column(name = "location")
   private String location;
-}
+
+
+  
+  }
+
